finish report system user and admin view

+++ admin/index.php
@@ -4,7 +4,19 @@
         include '../function.php';
         $name = $_SESSION['user'];
         $data = new Common();
-        $datas = $data->getAllData("SELECT * FROM report WHERE adminId = ?", [$_SESSION['id']]);
+        //  $datas = $data->getAllData("SELECT * FROM report WHERE adminId = ? ORDER BY date DESC", [$_SESSION['id']]);
+        $_SESSION['fromDate'] = $_POST['fromDate'];
+        $_SESSION['toDate'] = $_POST['toDate'];
+
+        require_once('../pagination/pagination_start.php');
+
+        $table = 'report'; 
+        $query = $data->getAllData("SELECT * FROM $table WHERE adminId = ? ORDER BY date DESC", [$_SESSION['id']]);
+        $result_count = count($query);
+        $total_no_of_pages = ceil($result_count / $total_records_per_page);
+        $second_last = $total_no_of_pages - 1;  
+        $datas = $data->getAllData("SELECT * FROM $table WHERE adminId = ? ORDER BY date DESC LIMIT $offset, $total_records_per_page", [$_SESSION['id']]);
+
     ?>
     
     <div class="container-fluid">
@@ -12,20 +24,24 @@
         <?php include 'top_menu.php' ?>
 
         <div class="d-flex justify-content-around mt-3">
-            <form class="example" action="action_page.php">
-                <!-- <input type="text" placeholder="Search.." name="search" class="input-search"> -->
-                <input type="date" name="fromDate" class="date-search" placeholder="From.....">
-                <input type="date" name="toDate" class="date-search" placeholder="To.....">
+            <form  method="post">
+                <input type="text" placeholder="Search.." name="keyword" class="input-search">
+                <label for="from">From</label>
+                <input type="date" name="fromDate" id="from" class="date-search" placeholder="From.....">
+                <label for="to">To</label>
+                <input type="date" name="toDate" id="to" class="date-search" placeholder="To.....">
                 <button type="submit" name="search" class="btn-submit"><i class="fa fa-search"></i></button>
             </form>
+
+            <div class="p-2 bg-info text-white">
+                <a href="adminExport.php" class="csv-link">Export Csv</a>
+            </div>
+
             <!-- add report -->
            <div class="btn btn-success report-new">
                <a href="new_report.php" class="text-white">Add New Report</a>
            </div>
 
-            <div class="p-2 bg-info text-white">
-                <a href="adminExport.php" class="csv-link">Export Csv</a>
-            </div>
         </div>
 
         <div class="row">
@@ -36,13 +52,10 @@
                 $keyword = $_POST['keyword'];
                 $fromDate = $_POST['fromDate'];
                 $toDate = $_POST['toDate'];
-
                 if(!empty($fromDate) && !empty($toDate)) {
                     $datas = $data->getAllData("SELECT * FROM report WHERE date between '".$fromDate."' and '".$toDate."' ");
-                }
-               
-                else {
-                    $datas = $data->getAllData("SELECT * FROM report WHERE date LIKE '%$keyword%' OR report LIKE '%$keyword%' ");
+                }else {
+                    $datas = $data->getAllData("SELECT * FROM report date LIKE '%$keyword%' OR report LIKE '%$keyword%' AND '".$id."' ");
                 }
             ?>
                 <table class="table table-striped mt-5">
@@ -58,7 +71,8 @@
                    <tbody>
                        <?php
                             if($datas) {
-                                foreach($datas as $row){ 
+                                foreach($datas as $row){
+                                    $i = 1;
                        ?>
                        <tr>
                             <td><?= $row['id'] ?></td>
@@ -67,16 +81,16 @@
                             <td><?= $row['report'] ?></td>
                             <td>
                                <a href="reportEdit.php?edit=<?= $row['id']?>" class="btn btn-outline-warning"><i class="fas fa-edit"></i></a> 
-                               <a href="reportDelete.php?delete=<?= $row['id']?>" class="btn btn-outline-danger"><i class="fas fa-trash-alt"></i></a>
+                               <a href="reportDelete.php?delete=<?= $row['id']?>" class="btn btn-outline-danger" onclick="return confirm('Are you sure to delete?')"><i class="fas fa-trash-alt"></i></a>
                             </td>
                        </tr>
                        <?php
                          }
-                            }
+                            $i++;}
                             else {
                        ?>
                         <tr>
-                            <td colspan="3" class="text-primary text-center font-weight-bold"><b>There is no datas to show</b></td>
+                            <td colspan="5" class="text-primary text-center font-weight-bold"><b>There is no datas to show</b></td>
                         </tr>
                         <?php } ?>
                    </tbody>
@@ -106,7 +120,7 @@
                             <td><?= $row['report'] ?></td>
                             <td>
                                <a href="reportEdit.php?edit=<?= $row['id']?>" class="btn btn-outline-warning"><i class="fas fa-edit"></i></a> 
-                               <a href="reportDelete.php?delete=<?= $row['id']?>" class="btn btn-outline-danger"><i class="fas fa-trash-alt"></i></a>
+                               <a href="reportDelete.php?delete=<?= $row['id']?>" class="btn btn-outline-danger" onclick="return confirm('Are you sure to delete?')"><i class="fas fa-trash-alt"></i></a>
                             </td>
                        </tr>
                        <?php
@@ -115,11 +129,21 @@
                                 else {
                        ?>
                        <tr>
-                            <td colspan="3" class="text-primary text-center font-weight-bold"><b>There is no datas to show</b></td>
+                            <td colspan="5" class="text-primary text-center font-weight-bold"><b>There is no datas to show</b></td>
                         </tr>
                         <?php } ?>
                    </tbody>
                 </table>
+                <!-- ====================== pagination ============================================ -->
+                <?php
+                    $serialize_user = serialize($user_arr);
+                ?>
+                <textarea name="export_data" id="" style="display:none;"><?= $serialize_user; ?></textarea>
+                <ul class="pagination">
+                    <?php
+                        include_once('../pagination/pagination_end.php');
+                    ?>
+                </ul>
                 <?php } ?>
             </div>
             <div class="col-md-1"></div>
